(REFACTORING) -> Product entity added

+++ src/main/java/com/example/gasManager/DTO/OrderDTO.java
@@ -2,6 +2,6 @@ package com.example.gasManager.DTO;
 
 import java.time.Instant;
 
-public record OrderDTO(Double totalValue, Instant saleDate, Long clientId) {
+public record OrderDTO(Double totalValue, Instant saleDate, Integer clientId) {
 
 }
